fix(portal): bump zustand to v5.0.3
This change bumps the zustand package
to version ^5.0.3. This is required to
fix the 'import.meta' is unsupported
error from blocking rendering in new
expo-managed react-native projects.

This is a major version change for the
library. There was some support dropped
for certain features. See the attached
migration guide for more details.

BREAKING CHANGE: major version bump for zustand dependency

refs: https://zustand.docs.pmnd.rs/migrations/migrating-to-v5

refs: #67